move processmanager -> processes/processmanager
