+++ processes/processmanager/process_options.go
@@ -17,7 +17,7 @@ package processmanager
 import (
 	"io"
 
-	"github.com/ligato/cn-infra/processmanager/status"
+	"github.com/ligato/cn-infra/processes/processmanager/status"
 )
 
 // POptions is common object which holds all selected options
