+++ processes/processmanager/status/status_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 
 	"github.com/ligato/cn-infra/logging/logrus"
-	"github.com/ligato/cn-infra/processmanager/status"
+	"github.com/ligato/cn-infra/processes/processmanager/status"
 	. "github.com/onsi/gomega"
 )
