+++ processes/processmanager/template/template.go
@@ -23,7 +23,7 @@ import (
 	"path/filepath"
 
 	"github.com/ligato/cn-infra/logging"
-	"github.com/ligato/cn-infra/processmanager/template/model/process"
+	"github.com/ligato/cn-infra/processes/processmanager/template/model/process"
 	"github.com/pkg/errors"
 )
